TEP-0056: Pipelines in Pipelines [Proposal]
Today, users can define and execute `Tasks` and `Custom Tasks`
in `Pipelines`. In this TEP, we propose allowing users to define
and execute `Pipelines` in `Pipelines`, alongside `Tasks` and
`Custom Tasks`.

We previously described the motivation for this TEP in:
- #374
- #498
- #698

In this change, we add an overview of the proposal which includes
the API changes to the specification and status of `Pipelines`.

In summary, we plan to support `PipelineRef` and `PipelineSpec`
only initially. The API design is extensible to support other fields
related to creating a `PipelineRun`.